Started to work on creating waves and spawning waves with OOB
I wan tto be able to create basic waves for now that i can spawn in different spawn location within a map. And the benefit of doing this with OOB makes it so that I can add as many waves as I want with all kinds of different types of enemys and with how many spawn point that I give it.
this will get more advance in the future. at the moment I can create a basic wave with a amount of enemys a enemytype and a list of spawn locations wich it will use to randomly spawn ech enemy at the spawnpoints within the list.